+++ src/index.ts
@@ -27,6 +27,22 @@ const openFileExplorer = () => {
   return isWsl;
 };
 
+const validatePath = (pathString: string) => {
+  const windowsPathPattern =
+    '(?:^[A-Za-z]:\\\\(?:[^<>:"/\\\\|?*]+\\\\)*[^<>:"/\\\\|?*]*$)';
+  const linuxPathPattern = "(?:^(?:\\/[a-zA-Z0-9_.-]+)*\\/?$)";
+  const regex = new RegExp(`${windowsPathPattern}|${linuxPathPattern}`);
+  return regex.test(pathString);
+  // What if path starts with C:\\?
+  // What if path has back-slashes?
+  // What if path has .. or .?
+  // What if path has back-slashes and forward-slashes? Throw error
+  // What if path has a substring with a space followed by no character? Throw error
+  // What if path has no slashes but has spaces Throw error
+};
+
+const sanitizePath = () => {};
+
 const batchRename: BatchRename = async (pathString) => {
   console.log("pathString:", pathString);
 };
@@ -37,21 +53,25 @@ const batchRenameCLI: BatchRename = async (pathString) => {
   console.log("files:", files);
   // Prompt user if they want to proceed or cancel
   // Loop through files asynchronously and cancel operation if an error occurs
-
-  // Option 1
-  // Copy file names
-  // Iterate and rename copied file names
-
-  // Option 2
-  // Iterate and rename file names
 };
 
 /* BRAINSTORMING
  * Renames files based on directory
  *
- * 1. Determine operating system
- *  Do I need to determine if OS is running in WSL?
- * 2. User selects a folder they want to rename files
+ *
+ * Option 1
+ * Copy file names
+ * Iterate and rename copied file na*
+ *
+ * Option 2
+ * Iterate and rename file names
+ *
+ * 1. Validate path
+ * 1a. Determine operating system
+ *    If os is win32 and
+ * 3. User selects a folder they want to rename files
  *  What if user selects files?
+ *
+ * How to dynamically get mount
  */
-export { batchRename, batchRenameCLI, openFileExplorer, isWSL };
+export { batchRename, batchRenameCLI, openFileExplorer, validatePath, isWSL };