+++ src/types/env.d.ts
@@ -1,5 +1,6 @@
 declare namespace NodeJS {
   interface ProcessEnv {
     WSL_WINDOWS_DIR: string;
+    WINDOWS_DIR: string;
   }
 }